Move ClientIdentifiers.java to 'lobby-client' so shared code can read it
This update consolidates http-client-lib stuff. This is so we can then
read that shared code from the lobby-server, which will allow us to
re-enable the error reporting endpoints
