+++ game-app/game-core/build.gradle
@@ -3,8 +3,6 @@ dependencies {
     implementation project(":game-app:map-data")
     implementation project(":game-app:game-relay-server")
     implementation project(":http-clients:lobby-client")
-    implementation project(":http-clients:maps-server")
-    implementation project(":lib:http-client-lib")
     implementation project(":lib:java-extras")
     implementation project(":lib:swing-lib")
     implementation project(":lib:websocket-client")
