+++ game-app/game-headed/build.gradle
@@ -19,9 +19,7 @@ dependencies {
     implementation project(":game-app:game-core")
     implementation project(":game-app:map-data")
     implementation project(":http-clients:lobby-client")
-    implementation project(":http-clients:maps-server")
     implementation project(":lib:feign-common")
-    implementation project(":lib:http-client-lib")
     implementation project(":lib:java-extras")
     implementation project(":lib:swing-lib")
     implementation project(":lib:websocket-client")
