+++ http-clients/game-support-server/build.gradle
@@ -2,7 +2,6 @@ dependencies {
     implementation "io.github.openfeign:feign-core:$feignVersion"
     implementation "io.github.openfeign:feign-gson:$feignVersion"
     implementation project(":lib:feign-common")
-    implementation project(":lib:http-client-lib")
     implementation project(":lib:java-extras")
     testImplementation "ru.lanwen.wiremock:wiremock-junit5:$wireMockJunit5Version"
     testImplementation project(":lib:test-common")
