+++ http-clients/lobby-client/src/main/java/org/triplea/http/client/ClientIdentifiers.java
@@ -1,4 +1,4 @@
-package org.triplea.http.client.lib;
+package org.triplea.http.client;
 
 import java.util.HashMap;
 import java.util.Map;
