+++ http-clients/lobby-client/src/main/java/org/triplea/http/client/maps/listing/MapDownloadItem.java
