+++ settings.gradle
@@ -9,9 +9,7 @@ include 'game-app:map-data'
 include 'game-app:smoke-testing'
 //include 'http-clients:game-support-server'
 include 'http-clients:lobby-client'
-include 'http-clients:maps-server'
 include 'lib:feign-common'
-include 'lib:http-client-lib'
 include 'lib:java-extras'
 include 'lib:swing-lib'
 include 'lib:swing-lib-test-support'
